+++ ruby/CHANGELOG.md
@@ -1,8 +1,9 @@
 ### Ruby images changelog
 
-#### Version ruby:2.x-3.x.x
+#### Version ruby:2.x-3.1.0
 
 * Fixed typo in `wait_for_node_install.sh`
+* Adds new entrypoint script to check for newer version of the base image
 
 #### Version ruby:2.x-3.0.3